+++ pmp-client/src/routes/parameters/ParameterList.tsx
@@ -14,7 +14,6 @@ const ParameterList = (props: ParameterListProps) => {
 	const { parameters, service } = props;
 	const [sortingCatagory, setSortingCatagory] = useState<SortingCatagory>(SortingCatagory.NAME);
 	const [sortingDirection, setSortingDirection] = useState(1);
-	const [filter] = useParameterFilter();
 
 	const sort = (parameters: Parameter[], catagory: SortingCatagory, sortingDirection: number) => {
 		if (catagory === SortingCatagory.NAME)
@@ -63,9 +62,9 @@ const ParameterList = (props: ParameterListProps) => {
 					</DataTableRow>
 				</DataTableHead>
 				<DataTableBody>
-					{sortedParameters.map((parameter) => (
-						(filter.types?.includes(parameter.type) && <ParameterListRow key={parameter.id} service={service} parameter={parameter} />)
-					))}
+					{sortedParameters.map((parameter) =>
+						<ParameterListRow key={parameter.id} service={service} parameter={parameter} />
+					)}
 				</DataTableBody>
 			</DataTableContent>
 		</DataTable>
