+++ pmp-client/src/routes/parameters/ParametersPage.tsx
@@ -2,9 +2,8 @@ import './style.css';
 
 import ServiceList from './ServiceList';
 import ParameterSearch from '../../features/search_filter/ParameterSearch';
-import { ParameterFilter } from '../../features/parameters/types';
-import { useState } from 'react';
 import ParameterFilterProvider from '../../features/search_filter/ParameterFilterProvider';
+import ParameterFilter from '../../features/search_filter/ParameterFilter';
 
 /**
  * Page for displaying and editing parameters.
@@ -15,6 +14,7 @@ const ParametersPage = () => {
 			<div className='paramPage'>
 				<h1>Parameters</h1>
 				<ParameterSearch/>
+				<ParameterFilter/>
 				<ServiceList />
 			</div>
 		</ParameterFilterProvider>
